requirements.txt
torch>=1.10.0
transformers>=4.15.0
psycopg2-binary>=2.9.3
pandas>=1.3.0
scikit-learn>=0.24.0
numpy>=1.19.0
matplotlib>=3.3.0
cx_Freeze>=6.10; platform_system=="Windows"
tkinter; platform_system=="Windows"
safetensors>=0.3.0 